Add configurable "sleep mode" setting #35
We currently have no certain way of knowing whether the device is battery powered, or USB mains adapter powered. The chip doesn't support it, and even if it did, the USB cable could be plugged into a USB power bank, making it battery powered still.

The device can measure the incoming voltage, but after it's been converted to the 3.7-4.2 range that the ESP32 needs, so we don't know if it's USB powered (5V) or battery powered (3.7-4.2V) either.

To decide whether or not to go to sleep, the device uses a clever heuristic: if the incoming voltage is very stable, it assumes it's on battery power. Otherwise, if the incoming voltage readings vary, it assumes it's on USB power.

This heuristic comes from measurements in the field, but it's not waterproof; a user reports that his USB power supply is so stable that the device thinks it's on battery power and goes to sleep while it could just as well stay awake. Even if we tweak it further, it will never be able to detect whether it's plugged into a USB power bank or a USB mains power adapter.

To resolve this, we propose a new configuration setting: config_sleep_mode

It can have 3 values:

no_sleep

    Ideal when plugged into mains power or using a solar setup with battery backup.
    Ensures instant transaction recognition and (if "always run webserver" is enabled) anytime configuration over wifi

custom_sleep

    Sleeps according to a configuration setting called custom_sleep_minutes with a default of 360 minutes (= 6 hours)
    Suitable for battery-powered setups, balancing power efficiency with functionality.

endless_sleep

    Only wakes when the action button is pressed or the tilt sensor is triggered.
    Perfect for maximizing battery life when continuous monitoring isn’t required.
    Requires manual wake-ups to check for new transactions.