Improve captive portal and configuration page login look and feel
When a user connects to the "Piggy Config" access point, their device
will automatically detect that it's a "Captive Portal" and redirect
them to the configuration page.

If they're not logged in yet, they will be presented with a good looking
form where they can do so. When they submit the form, they are forwarded
to the configuration page.